# Restaurante/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Restaurante",
    "description": "Clase: IS-513 Seccion 1700  Proyecto Grupo Restaurante. Este API maneja las operaciones de un restaurante._\n\nLinks:\n- [Repositorio del proyecto](https://github.com/AdrianJMendez/Restaurante)",
    "contact": {
      "email": "viktor.hernandez@unah.hn"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8091"
    }
  ],
...
  "paths": {
...
    "/api/restaurante/inventario/nombre/{nombre}": {
      "get": {
        "tags": [
          "Inventario"
        ],
        "summary": "Buscar el inventario por su nombre",
        "description": "Devuelve un inventario",
        "parameters": [
          {
            "name": "nombre",
            "in": "path",
            "description": "nombre del inventario",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion Exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventario"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/restaurante/inventario/crear": {
      "post": {
        "tags": [
          "Inventario"
        ],
        "summary": "Crear un registro del inventario",
        "description": "Crear un inventario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre del item del inventario",
                    "example": "Arroz"
                  },
                  "precio": {
                    "type": "integer",
                    "format": "int64",
                    "description": "Precio de compra del platillo",
                    "example": 15
                  },
                  "cantidad": {
                    "type": "number",
                    "description": "Cantidad del item que hay en existencia",
                    "example": 20
                  },
                  "unidadMetrica": {
                    "type": "string",
                    "description": "Unidad metrica",
                    "example": "libra"
                  },
                  "minimoRecompra": {
                    "type": "number",
                    "description": "Cantidad minima del item",
                    "example": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventario"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/restaurante/inventario/borrar/nombre/{nombre}": {
      "delete": {
        "tags": [
          "Inventario"
        ],
        "summary": "Borra el inventario por su nombre",
        "description": "Borra un inventario",
        "parameters": [
          {
            "name": "nombre",
            "in": "path",
            "description": "Nombre del inventario a eliminar",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion Exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Respuesta": {
                      "type": "string",
                      "example": "Inventario eliminado"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/restaurante/inventario/borrar/id/{inventarioId}": {
      "delete": {
        "tags": [
          "Inventario"
        ],
        "summary": "Borra el inventario por su id",
        "description": "Borra un inventario",
        "parameters": [
          {
            "name": "inventarioId",
            "in": "path",
            "description": "Id del inventario a eliminar",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion Exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Respuesta": {
                      "type": "string",
                      "example": "Inventario eliminado"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/restaurante/inventario/actualizar/id/{inventarioId}": {
      "put": {
        "tags": [
          "Inventario"
        ],
        "summary": "Actulizar un inventario",
        "description": "Actualizar un inventario",
        "parameters": [
          {
            "name": "inventarioId",
            "in": "path",
            "description": "Id del inventario a actualizar",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre del item del inventario",
                    "example": "Arroz"
                  },
                  "precio": {
                    "type": "integer",
                    "format": "int64",
                    "description": "Precio de compra del platillo",
                    "example": 15
                  },
                  "cantidad": {
                    "type": "number",
                    "description": "Cantidad del item que hay en existencia",
                    "example": 20
                  },
                  "unidadMetrica": {
                    "type": "string",
                    "description": "Unidad metrica",
                    "example": "libra"
                  },
                  "minimoRecompra": {
                    "type": "number",
                    "description": "Cantidad minima del item",
                    "example": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Platillo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/restaurante/inventario/actualizar/nombre/{nombre}": {
      "put": {
        "tags": [
          "Inventario"
        ],
        "summary": "Actulizar un inventario",
        "description": "Actualizar un inventario",
        "parameters": [
          {
            "name": "nombre",
            "in": "path",
            "description": "Nombre del inventario a actualizar",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre del item del inventario",
                    "example": "Arroz"
                  },
                  "precio": {
                    "type": "integer",
                    "format": "int64",
                    "description": "Precio de compra del platillo",
                    "example": 15
                  },
                  "cantidad": {
                    "type": "number",
                    "description": "Cantidad del item que hay en existencia",
                    "example": 20
                  },
                  "unidadMetrica": {
                    "type": "string",
                    "description": "Unidad metrica",
                    "example": "libra"
                  },
                  "minimoRecompra": {
                    "type": "number",
                    "description": "Cantidad minima del item",
                    "example": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventario"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
...
    "/api/restaurante/usuario/obtener/user/{user}": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "summary": "Buscar el usuario por su nombre",
        "description": "Devuelve un usuario",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "description": "Nombre del usuario",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion Exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/restaurante/usuario/crear": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "summary": "Crear un usuario",
        "description": "Crear un usuario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user": {
                    "type": "string",
                    "description": "Nombre del usuario",
                    "example": "Jose"
                  },
                  "categorioPermiso": {
                    "type": "object",
                    "properties": {
                      "categoriaPermiso": {
                        "type": "integer",
                        "description": "Categoria de permiso",
                        "example": 1
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/restaurante/usuario/borrar/id/{usuarioId}": {
      "delete": {
        "tags": [
          "Usuario"
        ],
        "summary": "Borra el usuario por su id",
        "description": "Borra un usuario",
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "description": "Id del usuario a eliminar",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion Exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Respuesta": {
                      "type": "string",
                      "example": "Usuario eliminado"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/restaurante/usuario/borrar/user/{user}": {
      "delete": {
        "tags": [
          "Usuario"
        ],
        "summary": "Borra el usuario por su nombre",
        "description": "Borra un usuario",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "description": "Nombre del usuario a eliminar",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion Exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Respuesta": {
                      "type": "string",
                      "example": "Usuario eliminado"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/restaurante/usuario/actualizar/id/{usuarioId}": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "summary": "Actulizar un usuario",
        "description": "Actualizar un usuario",
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "description": "Id del usuario a actualizar",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user": {
                    "type": "string",
                    "description": "Nombre del usuario",
                    "example": "Jose"
                  },
                  "categorioPermiso": {
                    "type": "object",
                    "properties": {
                      "categoriaPermiso": {
                        "type": "integer",
                        "description": "Categoria de permiso",
                        "example": 1
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Usuario": {
        "type": "object",
        "properties": {
          "usuarioId": {
            "type": "integer",
            "format": "int64",
            "description": "Generado por la base de datos",
            "example": 1
          },
          "user": {
            "type": "string",
            "description": "Nombre del usuario",
            "example": "Adrian"
          },
          "categorioPermiso": {
            "$ref": "#/components/schemas/CategoriaPermiso"
          }
        }
      },
      "CategoriaPermiso": {
        "type": "object",
        "properties": {
          "categoriaPermisoId": {
            "type": "integer",
            "format": "int64",
            "description": "Solo hay 3 tipos de categorias",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre de la categoria",
            "example": "Gerente"
          },
          "descripcion": {
            "type": "string",
            "example": "Puede crear nuevos usuarios"
          }
        }
      },
      "Inventario": {
        "type": "object",
        "properties": {
          "inventarioId": {
            "type": "integer",
            "format": "int64",
            "description": "Generado por la base de datos",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del item del inventario",
            "example": "arroz"
          },
          "precioCompra": {
            "type": "number",
            "description": "Precio de compra del item por unidad",
            "example": 10
          },
          "cantidad": {
            "type": "number",
            "description": "Cantidad del item que se encuentra en existencia",
            "example": 10
          },
          "unidadMetrica": {
            "type": "string",
            "description": "Unidad con la que se mide el item",
            "example": "libra"
          },
          "minimoRecompra": {
            "type": "number",
            "description": "Cantidad minima del item que debe haber en existencia",
            "example": 5
          }
        }
      },
      "Platillo": {
        "type": "object",
        "properties": {
          "platilloId": {
            "type": "integer",
            "format": "int64",
            "description": "Generado por la base de datos",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del platillo",
            "example": "Bistec"
          },
          "precio": {
            "type": "number",
            "description": "Precio de venta del platillo",
            "example": 100
          },
          "imagen": {
            "type": "string",
            "description": "URL de la imagen a mostrar en el frontend del platillo",
            "example": "bisctec.jpg"
          },
          "inventarios": {
            "type": "array",
            "description": "Items del inventario que lleva la receta del platillo",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Inventario"
                },
                {
                  "type": "integer",
                  "description": "Cantidad del item del inventario que lleva el platillo",
                  "example": "cantidad = 2"
                }
              ]
            }
          }
        }
      },
...
    }
  }
}